ASL-Smart-Glove-master/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
numpy==1.26.4
scikit-learn==1.2.2
requests==2.32.2
django-environ==0.11.2
whitenoise==6.6.0
psycopg2==2.9.7
pandas==2.2.2
google-generativeai==0.6.0
django-cors-headers==4.3.1
